Check for unimported Bijectors (#41)
Summary:
### Motivation
To make it easier for developers to add new Bijectors, we want a unit test that checks that all Bijectors that have been defined in the bijectors module to be imported by that module.

### Changes proposed
This PR defines a unit test that compares `flowtorch.bijectors.__all__` to a list of `Bijector` classes formed from walking the source files in `flowtorch.bijectors`. It raises an exception when a class defined in the latter is not imported by the former.

I think this is preferable to raising an exception on `import flowtorch.bijectors`.

I will do the auto-gen script for `flowtorch.bijectors/__init__.py` once we've merged in these PRs

Pull Request resolved: https://github.com/facebookincubator/flowtorch/pull/41

Reviewed By: horizon-blue

Differential Revision: D29276822

Pulled By: feynmanliang

fbshipit-source-id: 2796ccde6ee9466f4ef9c69f07d3218dbc7b53c5