Generate API documentation on-the-fly (#49)
Summary:
### Motivation
The API documentation is automatically generated from source. Therefore, we shouldn't keep a copy in the Git repo and generate it as required.

### Changes proposed
Additional step to "documentation" GitHub workflow to generate the API doc markdown files and JS sidebar, both for testing and before the final website is pushed to be live. Also, the generated files are added to .gitignore

Pull Request resolved: https://github.com/facebookincubator/flowtorch/pull/49

Reviewed By: jpchen

Differential Revision: D29684123

Pulled By: feynmanliang

fbshipit-source-id: 16d2e56c5457c42ebc17e50f18802298a0e429f0