Lazy parameters and bijectors with metaclasses (#59)
Summary:
feynmanliang my fork of `facebookincubator/flowtorch` was deleted when I connected my FB employee account to GitHub, so I have duplicated https://github.com/facebookincubator/flowtorch/issues/58 here (and added some additional functionality)

### Motivation
Shape information for a normalizing flow only becomes known when the base distribution has been specified. We have been searching for an ideal solution to express the delayed instantiation of Bijector and Params for this purpose. Several possible solutions are outlined in https://github.com/facebookincubator/flowtorch/issues/57.

### Changes proposed
The purpose of this PR is to showcase a prototype for a solution that uses metaclasses to express delayed instantiation. This works by intercepting .__call__ when a class initiated and returning a lazy wrapper around the class and bound arguments if only partial arguments are given to .__init__. If all arguments are given then the actual object is initialized. The lazy wrapper can have additional arguments bound to it, and will only become non-lazy when all the arguments are filled (or have defaults).

Pull Request resolved: https://github.com/facebookincubator/flowtorch/pull/59

Reviewed By: jpchen

Differential Revision: D30782184

Pulled By: stefanwebb

fbshipit-source-id: c53649d26e25478565681063361ad4e5a32110be