First sample scripts (#70)
Summary:
### Motivation
We would like a number of example scripts to demonstrate how to use FlowTorch.

### Changes proposed
I have created a new folder, `/samples`, and added the simple example from the landing page of the website. At the moment it is a Python script, although I think in the future they will be converted into Jupyter notebooks that are mirrored on Colab.

Pull Request resolved: https://github.com/facebookincubator/flowtorch/pull/70

Test Plan: The sample plots figures that demonstrate learning is working.

Reviewed By: horizon-blue

Differential Revision: D31580542

Pulled By: stefanwebb

fbshipit-source-id: dd594e085a89d90f7227361298b789a28a329a42