Added `.normalizing` method to `Flow` (#69)
Summary:
### Motivation
One of the advantages of normalizing flows is that you can take a sample from a distribution and transform it to Gaussian noise. This is useful as a stand alone operation for e.g. the NeuTra MCMC inference algorithm.

### Changes proposed
Each `flowtorch.distributions.Flow` object has a `.normalize` and `.rnormalize` method, analogous to `.sample` and `.rsample`. Normalize and sample evaluate the normalizing and generative directions of the flow, respectively.

Pull Request resolved: https://github.com/facebookincubator/flowtorch/pull/69

Test Plan: Unit tests to follow in the future. I will also post a demonstration script on a learnt bivariate normal distribution in a future PR.

Reviewed By: jpchen

Differential Revision: D31580523

Pulled By: stefanwebb

fbshipit-source-id: c66a5587c15814f9d4f0e1cee9c8b52ffe92c381