Autogenerating imports for for `flowtorch.parameters` and `flowtorch.bijectors` (#72)
Summary:
### Motivation
It is tiresome to have to add new components to __init__.py for bijectors, distributions and parameters. We should be able to automatically generate it!

### Changes proposed
Autogen for distributions was completed in a previous PR. This one completes it for parameters and bijectors.

I also uncovered and fixed a bug in how `utils.list_bijectors()`, `utils.list_distributions()`, and `utils.list_parameters` were working

Pull Request resolved: https://github.com/facebookincubator/flowtorch/pull/72

Reviewed By: horizon-blue

Differential Revision: D31602898

Pulled By: stefanwebb

fbshipit-source-id: 6aef290c8538492ee9feae5a7d252ee612eeade5