Bijective tensors for caching intermediate values (#1334)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1334

### Motivation
As described in https://github.com/facebookincubator/flowtorch/issues/88, we'd wish to have a way of caching intermediate values computed in the flow.

### Changes proposed
This PR assigns this responsibility to a new class, BijectiveTensor.
A BijectiveTensor keeps track of the layer that has created it, the original tensor and whether it comes from a call to 'forward' or 'inverse'.
It inherits from torch.Tensor. By default, an operation on a BijectiveTensor returns a torch.Tensor (except if this operation is a Bijector).
One can control if BijectiveTensors should be used (which is the case by default) with the context manager `set_record_flow_graph`.

Pull Request resolved: https://github.com/facebookincubator/flowtorch/pull/89

Test Plan:
A test file can be found in `test/test_bijectivetensor.py`.

### Types of changes
- [ ] Docs change / refactoring / dependency upgrade
- [ ] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

### Checklist
- [X] My code follows the code style of this project.
- [X] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [X] I have read the **[CONTRIBUTING](https://github.com/facebookincubator/flowtorch/blob/main/CONTRIBUTING.md)** document.
- [X] I have added tests to cover my changes.
- [X] All new and existing tests passed.
- [X] The title of my pull request is a short description of the requested changes.

Reviewed By: ToddSmall

Differential Revision: D33927956

Pulled By: stefanwebb

fbshipit-source-id: 2f3e53efb69f5839cb0649bc0d834cc7a0503553