Fixed bug in `bijectors.ops.spline.Spline` and unit test for log(detJ)) (#100)
Summary:
### Motivation
From other work, I discovered that training Neural Spline Flows was not working as expected, being unable to learn simple toy distributions... This PR fixes this, as well as the reason that the unit tests were not picking it up.

### Changes proposed
I changed the sign of the `log(det(J))` in the `inverse` method of `bijectors.ops.spline.Spline`, and ensured the unit tests are not using cached values of `log(det(J))` (via `BijectiveTensor`) when comparing `log(det(J))` from the forward method to that of the inverse one.

Pull Request resolved: https://github.com/facebookincubator/flowtorch/pull/100

Test Plan: Run `pytest tests/` and try the Neural Spline Flow example in the theory tutorials

Reviewed By: ToddSmall

Differential Revision: D34616134

Pulled By: stefanwebb

fbshipit-source-id: d29a94fd80b2e32a920da194a3c75c05c382e461